requirement.txt
aioredis==1.1.0
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.7.1
Automat==0.7.0
channels==2.1.2
channels-redis==2.2.1
constantly==15.1.0
daphne==2.2.1
Django==2.1
djangorestframework==3.9.2
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
msgpack==0.5.6
PyHamcrest==1.9.0
pytz==2018.5
selenium==3.14.0
six==1.11.0
Twisted==18.7.0
txaio==18.7.1
urllib3==1.23
zope.interface==4.5.0
django-filter==2.1.0
